% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{api_error}
\alias{api_error}
\title{Catch API errors}
\usage{
api_error(req)
}
\arguments{
\item{req}{Response from the API call}
}
\description{
Function checks the API status code
}
\details{
Based on the API status code, the function evaluates
if the response is valid or not. If not, a message is
displayed
}
