% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_location.R
\name{append_location}
\alias{append_location}
\title{Add location to dataframe}
\usage{
append_location(finishedDF, rawdat = rawdat)
}
\arguments{
\item{df}{final dataframe with the aggregated data}

\item{rawdat}{locations as text not EIC}
}
\value{
dataframe
}
\description{
Function calls the `get_Locations` function to get the country/location name
}
\details{
Adds a column to the finished dataframe, appending the corresponding locations
}
