% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areaY}
\alias{areaY}
\title{Area and EIC codes}
\format{
A data frame with 1213 rows and 11 columns:
\describe{
  \item{EicCode}{The EIC is based on fixed length alphanumeric codes which can be broken down as follows:
  * A 2-character number identifying the Issuing Office assigned by ENTSO-E.
  * One Character identifying the object type that the code represents.
  * 12 digits, uppercase characters or minus signs allocated by the issuing office
  * 1 check character to ensure the code validity.}
  \item{EicDisplayName}{Machine friendly name for Country/Region/Participant.}
  \item{EicLongName}{Name of participant.}
  \item{EicParent}{If two different EIC Participants have the same VAT
  identification number or identification code and this is permitted by the
  local tax regulations, one of the two EIC codes should be designated as
  EIC Parent, otherwise one of the two EIC codes will not be accepted.}
  \item{EicResponsibleParty}{Responsible party for the publication.}
  \item{EicStatus}{Status of the EIC code (active/inactive).}
  \item{MarketParticipantPostalCode}{Postal code of the participating Entity.}
  \item{MarketParticipantIsoCountryCode}{Country code of the participating Entity.}
  \item{MarketParticipantVatCode}{VAT code of the participating Entity.}
  \item{EicTypeFunctionList}{}
  \item{type}{}
}
}
\source{
<https://eepublicdownloads.entsoe.eu/eic-codes-csv/Y_eiccodes.csv>
}
\usage{
areaY
}
\description{
Dataset pulled directly from the Entso-E website
including all the available EIC and Area codes
for each country/region etc.
Areas for inter System Operator data interchange.
}
\keyword{datasets}
