% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_url.R
\name{build_url}
\alias{build_url}
\title{Build URL}
\usage{
build_url(
  documentType,
  processType,
  contract_MarketAgreement.Type,
  outBiddingZone_Domain,
  auction.Type,
  psrType,
  Area_Domain,
  in_Domain,
  out_Domain,
  registeredResource,
  businessType,
  periodStart,
  periodEnd,
  key = TRUE
)
}
\arguments{
\item{documentType}{Corresponding to the data point}

\item{processType}{Which process should be pulled}

\item{contract_MarketAgreement.Type}{Type of Contract}

\item{outBiddingZone_Domain}{Country from which the data should be pulled}

\item{auction.Type}{Type of auction, buy or sell}

\item{psrType}{Type of generation, e.g. wind, solar, nuclear etc.}

\item{Area_Domain}{Domain Area}

\item{in_Domain}{From where}

\item{out_Domain}{To where}

\item{registeredResource}{Which generation unit}

\item{businessType}{Identification of the nature of time series}

\item{periodStart}{Starting point}

\item{periodEnd}{End point}

\item{key}{Should the API-Key be included (mainly for testing purpose)?}
}
\value{
String of url to make request
}
\description{
Dynamically build the right URL for your request.
}
\details{
This function builds the url to get the data.
Different types of data have different arguments in the function.
With that in mind, arguments can be missing based on the
desired data.
}
