% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeList}
\alias{codeList}
\title{Mulitple datapoints and code descriptions for Entso-E API}
\format{
A list containing 27 different dataframes
\describe{
 \item{UnitOfMeasureTypeList}{Unit of supplied data}
 \item{TarifTypeTypeList}{Tarif type}
 \item{StatusTypeList}{Status of the document}
 \item{RoleTypeList}{Role of the party}
 \item{RightsTypeList}{Type of right of the data point}
 \item{ReasonCodeTypeList}{Indicating the reason for acceptance or rejection}
 \item{QualityTypeList}{Quality of the data}
 \item{ProcessTypeList}{Process type, information provided and on which
 schedule}
 \item{PriceDirectionTypeList}{Income or expediture (sell or buy)}
 \item{PriceCategoryTypeList}{Price categorie}
 \item{PaymentTermsTypeList}{Type of payment}
 \item{ObjectAggregationTypeList}{Type of object, area, party etc.}
 \item{IndicatorTypeList}{Indicator}
 \item{EicTypeList}{Type of EIC identifier}
 \item{EnergyProductTypeList}{Active or reactive power/energy}
 \item{DocumentTypeList}{Requested document type}
 \item{DirectionTypeList}{Increase or decrease energy}
 \item{CurveTypeList}{Curve type}
 \item{CurrencyTypeList}{Currency}
 \item{ContractTypeList}{Allocation of capacity}
 \item{CodingSchemeTypeList}{Type of coding}
 \item{ClassificationTypeList}{Classification}
 \item{CategoryTypeList}{Category type}
 \item{BusinessTypeList}{Type of business of the supplier}
 \item{AuctionTypeList}{Auction type}
 \item{AssetTypeList}{Type of generation/load}
 \item{AllocationModeTypeList}{Type of allocation to buy power}
}
}
\source{
<https://docstore.entsoe.eu/Documents/EDI/Library/Core/entso-e-code-list-v29r0.pdf>
}
\usage{
codeList
}
\description{
Dataset pulled from the Entso-E, transformed the pdf to list.
It includes many explenations for codes such as Document Type,
Business Type and many more
}
\keyword{datasets}
