% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_response.R
\name{date_from_lst}
\alias{date_from_lst}
\title{Apply date function to `TimeSeries` list}
\usage{
date_from_lst(lst_ts)
}
\arguments{
\item{lst_ts}{TimeSeries object from the function only_ts}
}
\value{
Returns a list of the date sequence, depending on the number
of columns the API response had
}
\description{
Function returns date-list from API response containing
only `TimeSeries` objects.
}
\details{
Function uses the points_to_time function to create the date sequence
It is needed to extract the different variables from the API response
}
