% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_colnames.R
\name{dynamic_colnames}
\alias{dynamic_colnames}
\title{Create dynamic colnames}
\usage{
dynamic_colnames(df, rawdat, onlyTS_dat, codeList, PSR = FALSE)
}
\arguments{
\item{df}{final dataframe (without proper colnames)}

\item{rawdat}{Converted XML response to list}

\item{onlyTS_dat}{list of data containing only the timeseries data}

\item{codeList}{codeList (excel data)}
}
\value{
vector with correct colnames
}
\description{
Function dynamically craetes colnames
}
\details{
Append the processtype (e.g. intraday total) to the colnames. This
is done with the `get_params` function.
}
