% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eicLoc}
\alias{eicLoc}
\title{EIC Area Code List}
\format{
A data frame with 53380 rows and 11 columns:
\describe{
  \item{mRID}{The unique identification of the time
  * series within the document.}
  \item{docStatus}{Only used in case a document has
  * been submitted by mistake. A13 = Withdrawn.}
  \item{Instance}{Not anymore used.}
  \item{Name}{The full name associated to the EIC code.}
  \item{Name_Abbr}{The display name or short name to be used on displays.}
  \item{Date_Change}{Date of the request.}
  \item{Date_Deactivate}{Date when the deactivation will becarried out.}
  \item{Street}{Street of the EIC Location. At least the attribute “country” shall be published.}
  \item{aCERCode}{The ACER code associated to the EIC code of the market participant.}
  \item{Description}{The description of the EIC code.}
  \item{Function_Name}{The function(s) of the EIC code.}
}
}
\source{
<https://eepublicdownloads.blob.core.windows.net/public-cdn-container/clean-documents/fileadmin/user_upload/edi/library/eic/allocated-eic-codes.xml>
<https://eepublicdownloads.entsoe.eu/clean-documents/EDI/Library/cim_based/EIC_Data_Exchange_IG_v1.1.pdf>
}
\usage{
eicLoc
}
\description{
Dateset pulled from the Entso-E server.
The data serves as identification table for the
locations of generation units. The df includes the name,
if given the street name and more. The dataset is primarely
used to append the location to a unit.
}
\keyword{datasets}
