% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_colnames.R
\name{generation_consumption}
\alias{generation_consumption}
\title{Add "Generation" and/or "Consumption" to colnames}
\usage{
generation_consumption(ts_dat)
}
\arguments{
\item{ts_dat}{Only_ts data}
}
\value{
vector with unique values
}
\description{
Function searches for the name "BiddingZone_Domain.mRID" and returns
if the data point is generation or consumption
}
\details{
For each "TimeSeries" object the function evaluates if the parameter
`outBiddingZone_Domain.mRID` or `inBiddingZone_Domain.mRID` is present.
This needs to be done to determine for some data points if the returned
data is for generation or consumption. The function also checks if the last
word of the colnames already is generation or consumption
}
