% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_location.R
\name{get_Locations}
\alias{get_Locations}
\title{Get location from mRID}
\usage{
get_Locations(rawdat, eicLoc = eicLoc)
}
\arguments{
\item{rawdat}{Converted XML response to list}

\item{eicLoc}{}
}
\value{
returns a vector with the locations
}
\description{
Function takes the mRID from the pulled data and matches
it with the `eicLoc` dataframe to get the real name
}
\details{
Takes all the unique Resource mRIDs from the raw data
and matches them to the eicLoc table, extracting the name
of the power plan
}
