% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_parameters.R
\name{get_params}
\alias{get_params}
\title{Get parameters}
\usage{
get_params(rlst)
}
\arguments{
\item{rlst}{Raw list, converted XML response to list}
}
\value{
named list with the data
}
\description{
Pull the different parameters from the original object.
}
\details{
Function extracts the Process Type, Business type,
object aggregation, Measurement unit and curve type.
}
