% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_response.R
\name{only_ts}
\alias{only_ts}
\title{Return only "TimeSeries" objects}
\usage{
only_ts(rlst)
}
\arguments{
\item{rlst}{Raw list, converted xml response}
}
\value{
Only the lists which are named TimeSeries
}
\description{
Function which only return "TimeSeries" objects
}
\details{
The response from the Entso-E API has a lot of
different elements called "TimeSeries". These elements
inlcude all the necessary information like point and
value. This function extracts all the elements with the
name "TimeSeries".
}
