% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_response.R
\name{points_to_time}
\alias{points_to_time}
\title{Get the date from point}
\usage{
points_to_time(start, n_points, period)
}
\arguments{
\item{start}{StartInterval from the API response}

\item{n_points}{Number of datapoints}

\item{period}{Periodicity of the response}
}
\value{
Date sequence from start with length.out = n_points
and period = periodicity
}
\description{
Function converts points to date&time
}
\details{
Function uses the `start` data point which is returned
by the API as a refference point. From that, it creates a sequence
with the periodicity of the response as well as the number of
points returned by the call. This sequence represents the dates
in the response.
}
