% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pull_data}
\alias{pull_data}
\title{Pull data from the API}
\usage{
pull_data(
  documentType,
  processType,
  contract_MarketAgreement.Type,
  outBiddingZone_Domain,
  auction.Type,
  psrType,
  Area_Domain,
  in_Domain,
  out_Domain,
  registeredResource,
  businessType,
  periodStart,
  periodEnd
)
}
\arguments{
\item{documentType}{Corresponding to the data point}

\item{processType}{Which process should be pulled}

\item{contract_MarketAgreement.Type}{Type of Contract}

\item{outBiddingZone_Domain}{Country from which the data should be pulled}

\item{auction.Type}{Type of auction, buy or sell}

\item{psrType}{Tpy of generation, e.g. wind, solar, nuclear etc.}

\item{Area_Domain}{Domain Area}

\item{in_Domain}{From where}

\item{out_Domain}{To where}

\item{registeredResource}{Which generation unit}

\item{businessType}{Identification of the nature of time series}

\item{periodStart}{Starting point}

\item{periodEnd}{End point}
}
\value{
Returns the finished data frame
}
\description{
Function to pull the data from the Entso-E API. The function
combines all functions from the package to be as dynamic
as possible. It handles all implemented data points (see github).
}
\details{
Function uses each element of this package. It build the url, pull the data and
converts it in the appropriate manner. The function is complex because different
types of data points need different treatment.
}
