% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{set_apikey}
\alias{set_apikey}
\title{Set your API key}
\usage{
set_apikey(key)
}
\arguments{
\item{key}{This is the ENSTO-E API Key}
}
\description{
`set_apikey` sets your API key to the environment variable
}
\details{
This function sets the API Key to the environment
By setting this key, the functions to pull the data
can rely on the environment rather than passing the key
each time. The key is stored in the `ENTSOE_KEY` environment
variable. It also checks if you already set a key.
}
