% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{split_columns}
\alias{split_columns}
\title{Split columns based on parameters}
\usage{
split_columns(df = temp_df, rawdat = converted_response)
}
\arguments{
\item{df}{Half-finished dataframe}

\item{rawdat}{Raw data, as list}
}
\value{
Split dataframe
}
\description{
Function converts long to wide dataframe based on parameters.
}
\details{
With the `pivot_wider` function the data is put into wide format
based on the parameters. Function is mainly used if the requested
data has two columns (i.e. scheduled generation and scheduled consumption)
}
