% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{to_daily}
\alias{to_daily}
\title{Convert to Daily}
\usage{
to_daily(df, type = c("mean", "median", "sum"))
}
\arguments{
\item{df}{Finished data frame returned from `pull_data`}

\item{type}{Type of aggregation. One can choose between `mean`, `median` and
`sum`}
}
\description{
Function takes the finished data frame and converts it to a daily
ts. There are three different aggregation methods.
}
\details{
The `type` argument specifies the type of the aggregation.
With the date column, the data is being aggregated.
}
