% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{to_yearly}
\alias{to_yearly}
\title{Convert to yearly}
\usage{
to_yearly(df, type = c("mean", "median", "sum"))
}
\arguments{
\item{df}{Finished data frame returned from `pull_data`}

\item{type}{Method of aggregation. One can choose between `mean`, `median` and
`sum`.}
}
\value{
Converted data frame
}
\description{
Function takes the finished data frame and converts it to a yearly
ts. There are three different aggregation methods.
}
\details{
A new column is added (year) derived from the date column.
The `type` argument specifies the type of the aggregation.
With the added columns, a aggregation is made.
}
