% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_parameters.R
\name{type_from_list}
\alias{type_from_list}
\title{PSR Type from Code}
\usage{
type_from_list(rlst, psrtype)
}
\arguments{
\item{rlst}{Rlst raw list, converted xml response to list}

\item{psrtype}{PSR Type from codeList$AssetTypeList (excel data)}
}
\value{
Returns named vector with the PSR types in order of them being
in the data
}
\description{
Extract the PSR Code and convert code to name
}
\details{
Function extracts the PSR Type (and Text) and generates a list
Is used to append to data and know what type it is (e.g. [14.1.D])
}
