% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_response.R
\name{values_from_lst}
\alias{values_from_lst}
\title{Get values from response}
\usage{
values_from_lst(lst_ts)
}
\arguments{
\item{lst_ts}{TimeSeries object from the function only_ts}
}
\value{
List with only values
}
\description{
Using the `only_ts` function, this function returns only the values.
}
\details{
Function pulls all the values from the response list (output from
`only_ts`). These values have the name point and the function returns
a list with the values
}
