% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_location.R
\name{append_location}
\alias{append_location}
\title{Add location to data frame}
\usage{
append_location(df, rawdat)
}
\arguments{
\item{df}{Final data frame with the aggregated data}

\item{rawdat}{Locations as text not EIC}
}
\value{
Data frame
}
\description{
Function calls the `get_Locations` function to get the country/location name.
}
\details{
Adds a column to the finished data frame, appending the corresponding locations.
}
