% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_url.R
\name{build_url}
\alias{build_url}
\title{Build URL}
\usage{
build_url(documentType = "A65", processType = "A16",
     outBiddingZone_Domain = "10YCZ-CEPS-----N",
     periodStart = "201512312300", periodEnd = "201612312300")
}
\arguments{
\item{documentType}{Corresponding to the data point}

\item{processType}{Which process should be pulled}

\item{contract_MarketAgreement.Type}{Type of Contract}

\item{outBiddingZone_Domain}{Country from which the data should be pulled}

\item{auction.Type}{Type of auction, buy or sell}

\item{psrType}{Tpy of generation, e.g. wind, solar, nuclear etc.}

\item{Area_Domain}{Domain Area}

\item{in_Domain}{From where}

\item{out_Domain}{To where}

\item{registeredResource}{Which generation unit}

\item{businessType}{Identification of the nature of timeseries}

\item{periodStart}{Starting point}

\item{periodEnd}{End point}
}
\value{
String of url to make request
}
\description{
Dynamically build the right URL for your request
}
\details{
This function builds the url to get the data.
Different types of data have different arguments in the function.
With that in mind, arguments can be missing based on the
desired data.
}
