% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_colnames.R
\name{business_types}
\alias{business_types}
\title{Get Business Types}
\usage{
business_types(ts_dat)
}
\arguments{
\item{ts_dat}{Response converted to onlyTS}
}
\value{
Vector of business types
}
\description{
Function gets the businesstype from raw data.
}
\details{
Function is subject to change. Lifecycle is experimental.
}
