% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convert_xml}
\alias{convert_xml}
\title{Convert XML to list}
\usage{
convert_xml(xml)
}
\arguments{
\item{xml}{Response object from the api call}
}
\value{
list of response
}
\description{
`convert_xml` returns a list from a xml response
}
\details{
This function converts the xml response to a list
From that list, the timeInterval, the period and
the raw data is extracted and returned as a list
with three sublists
}
