% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_colnames.R
\name{get_displayName}
\alias{get_displayName}
\title{Function pulls the name of the location from the eic code}
\usage{
get_displayName(rlst, eic)
}
\arguments{
\item{rlst}{Rlst raw list, converted xml response to list}

\item{eic}{Dataframe with the area codes (areaY from the entso e website)}
}
\value{
Short display name of the country/region/area
}
\description{
Function pulls the name of the location from the eic code
}
