% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pull_data}
\alias{pull_data}
\title{Pull data from the API}
\usage{
Actual Total Load [6.1.A]:
pull_data(
 documentType = "A65", processType = "A16",
 outBiddingZone_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Day-Ahead Total Load Forecast [6.1.B]:
pull_data(
 documentType = "A65", processType = "A01",
 outBiddingZone_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Week-Ahead Total Load Forecast [6.1.C]:
pull_data(
 documentType = "A65", processType = "A31",
 outBiddingZone_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Month-Ahead Total Load Forecast [6.1.D]:
pull_data(
 documentType = "A65", processType = "A32",
 outBiddingZone_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Year-Ahead Total Load Forecast [6.1.E]:
pull_data(
 documentType = "A65", processType = "A33",
 outBiddingZone_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Year-Ahead Forecast Margin [8.1]:
pull_data(
 documentType = "A70", processType = "A33",
 outBiddingZone_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Forecasted Capacity [11.1.A]:
pull_data(
 documentType = "A61", contract_MarketAgreement.Type = "A01",
 in_Domain = "10YCZ-CEPS-----N", out_Domain = "10YSK-SEPS-----K",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Installed Generation Capacity Aggregated [14.1.A]:
pull_data(
 documentType = "A68", processType = "A01",
 in_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Installed Generation Capacity per Unit [14.1.B]:
pull_data(
 documentType = "A71", processType = "A33",
 in_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Day-ahead Aggregated Generation [14.1.C]:
pull_data(
 documentType = "A71", processType = "A01",
 in_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Day-ahead Generation Forecasts for Wind and Solar [14.1.D]:
pull_data(
 documentType = "A68", processType = "A33",
 in_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Current Generation Forecasts for Wind and Solar [14.1.D]:
pull_data(
 documentType = "A69", processType = "A18",
 in_Domain = "10YFI-1--------U",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Intraday Generation Forecasts for Wind and Solar [14.1.D]:
pull_data(
 documentType = "A69", processType = "A40",
 in_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Actual Generation Output per Generation Unit [16.1.A]:
pull_data(
 documentType = "A73", processType = "A16",
 in_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300",
 registered_resource = "27W-GU-EDETG2--X" (optional!)
)

Aggregated Filling Rate of Water Reservoirs and Hydro Storage Plants [16.1.D]:
pull_data(
 documentType = "A72", processType = "A16",
 in_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)

Aggregated Generation per Type [16.1.B&C]:
pull_data(
 documentType = "A75", processType = "A16",
 in_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"
)
}
\arguments{
\item{documentType}{Corresponding to the data point}

\item{processType}{Which process should be pulled}

\item{contract_MarketAgreement.Type}{Type of Contract}

\item{outBiddingZone_Domain}{Country from which the data should be pulled}

\item{auction.Type}{Type of auction, buy or sell}

\item{psrType}{Tpy of generation, e.g. wind, solar, nuclear etc.}

\item{Area_Domain}{Domain Area}

\item{in_Domain}{From where}

\item{out_Domain}{To where}

\item{registeredResource}{Which generation unit}

\item{businessType}{Identification of the nature of timeseries}

\item{periodStart}{Starting point}

\item{periodEnd}{End point}
}
\value{
Dataframe
}
\description{
Function to pull the data from the Entso-E API.
}
\details{
Function uses each element of this package. It build the url, pull the data and
converts it in the appropriate manner. The funciton is complex because different
types of datapoints need different treatment.
}
