% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{to_daily}
\alias{to_daily}
\title{Convert to Daily}
\usage{
to_daily(pull_data(
 documentType = "A65", processType = "A16",
 outBiddingZone_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"), "median")
}
\arguments{
\item{df}{Dataframe pulled from the API}

\item{type}{Type of aggregation. One can choose between `mean`, `median` and
`sum`}
}
\description{
Function takes the pulled data (dataframe) and converts it to daily
}
\details{
The `type` argument specifies the type of the aggregation.
With the date column, the data is being aggregated.
}
