% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{to_monthly}
\alias{to_monthly}
\title{Convert to Monthly}
\usage{
to_monthly(df, type = c("mean", "median", "sum"))
}
\arguments{
\item{df}{Dataframe pulled from the API}

\item{type}{Type of aggregation. One can choose between `mean`, `median` and
`sum`}
}
\description{
Function takes the pulled data (dataframe) and converts it to monthly.
}
\details{
Two new columns are added (year, month) derived from the date column.
The `type` argument specifies the type of the aggregation.
With the added columns, a aggregation is made.
}
