% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{to_yearly}
\alias{to_yearly}
\title{Convert to yealy}
\usage{
to_yearly(pull_data(
 documentType = "A65", processType = "A16",
 outBiddingZone_Domain = "10YCZ-CEPS-----N",
 periodStart = "201912312300", periodEnd = "202012312300"), "median")
}
\arguments{
\item{df}{Dataframe pulled from the API}

\item{type}{Type of aggregation. One can choose between `mean`, `median` and
`sum`.}
}
\value{
Converted dataframe
}
\description{
Function takes tha pulled data (dataframe) and converts it to yearly.
}
\details{
A new column is added (year) derived from the date column.
The `type` argument specifies the type of the aggregation.
With the added columns, a aggregation is made.
}
